Styling

+++ components/ServiceProvider/index.js
@@ -24,23 +24,26 @@ const EditButton = styled.button`
   background-color: #3498db;
   color: white;
   border: none;
-  padding: 8px 16px;
+  padding: 5px;
   border-radius: 4px;
   cursor: pointer;
-  margin-top: 10px;
+  margin: 6px;
 `;
 
-const ActionButton = styled.button`
+const ShowReviewButton = styled.button`
   border: none;
   padding: 8px 16px;
   border-radius: 4px;
   cursor: pointer;
   margin-top: 10px;
   margin-right: 10px;
+  margin-bottom: 10px;
+  background-color: black;
+  color: white;
 `;
 
-const ReviewButton = styled(ActionButton)`
-  background-color: #2ecc71;
+const ReviewButton = styled(ShowReviewButton)`
+  background-color:  #FF5733;
   color: white;
 `;
 
@@ -48,10 +51,10 @@ const DeleteButton = styled.button`
   background-color: #e74c3c;
   color: white;
   border: none;
-  padding: 8px 16px;
+  padding: 5px;
   border-radius: 4px;
   cursor: pointer;
-  margin: 10px;
+  margin: 6px;
 `;
 
 const ShowContactButton = styled.button`
@@ -90,6 +93,13 @@ export default function ServiceProvider({ card, isOnFavoritesPage }) {
 
   async function handleDelete(id) {
     const url = `/api/providers/${id}`;
+
+    const userConfirmed = window.confirm(
+      "Do you really want to permanently delete this card? You cannot restore the card after confirmation!"
+    );
+    if (!userConfirmed) {
+      return;
+    }
     try {
       const response = await fetch(url, {
         method: "DELETE",
@@ -99,7 +109,7 @@ export default function ServiceProvider({ card, isOnFavoritesPage }) {
         await response.json();
 
         mutate();
-        alert("You have successfully Deleted!");
+        alert("You have successfully deleted the card!");
       } else {
         console.error(`Error: ${response.status}`);
       }
@@ -121,26 +131,53 @@ export default function ServiceProvider({ card, isOnFavoritesPage }) {
           <h2>
             {card.firstName} {card.lastName}
           </h2>
-          <h3>
+          <p>
             <strong>Skills:</strong> {card.skills}
-          </h3>
-          <h3>
+          </p>
+          <p>
             <strong>Needs:</strong> {card.needs}
-          </h3>
+          </p>
           {showContactInfo && (
             <div>
-              <h3>
+              <p>
                 <strong>Email:</strong> {card.email}
-              </h3>
-              <h3>
+              </p>
+              <p>
                 <strong>Phone:</strong> {card.phone}
-              </h3>
+              </p>
             </div>
           )}
           <ShowContactButton type="button" onClick={toggleContactInfo}>
             {showContactInfo ? "Hide Contact" : "Show Contact"}
           </ShowContactButton>
+            <hr></hr>
+            <details>
+              <summary>Give me a Review</summary> 
+          <ReviewButton onClick={toggleReviewForm}>
+            {showReviewForm ? "Hide Review Form" : "Add Review"}
+          </ReviewButton>
 
+          {showReviewForm && <ReviewForm card={card} />}
+
+          <ShowReviewButton onClick={toggleReviews}>
+            {showReviews ? "Hide Reviews" : "Show Reviews"}
+          </ShowReviewButton>
+          </details>
+          {showReviews && card.reviews && (
+            <article>
+              <h2>Reviews:</h2>
+              {card.reviews.map((review, id) => (
+                <h3 key={id}>{review}</h3>
+              ))}
+            </article>
+          )}
+             <hr></hr>
+          <details>
+            <summary>Give me a Rating</summary>
+            <p>You are welcome to rate my service here. Thank you very much!</p> 
+            <StarRating card={card} />
+          </details>
+          <hr></hr>  
           {!isOnFavoritesPage && (
             <EditDeleteWrapper>
               <EditButton type="button" onClick={handleOpenEditForm}>
@@ -154,28 +191,6 @@ export default function ServiceProvider({ card, isOnFavoritesPage }) {
               </DeleteButton>
             </EditDeleteWrapper>
           )}
-
-          <ReviewButton onClick={toggleReviewForm}>
-            {showReviewForm ? "Hide Review Form" : "Add Review"}
-          </ReviewButton>
-
-          {showReviewForm && <ReviewForm card={card} />}
-
-          <ActionButton onClick={toggleReviews}>
-            {showReviews ? "Hide Reviews" : "Show Reviews"}
-          </ActionButton>
-
-          {showReviews && card.reviews && (
-            <article>
-              <h2>Reviews:</h2>
-              {card.reviews.map((review, id) => (
-                <h3 key={id}>{review}</h3>
-              ))}
-            </article>
-          )}
-          <div>
-            <StarRating card={card} />
-          </div>
         </div>
       )}
     </ServiceProviderWrapper>
